+++ CHANGES.md
@@ -1,7 +1,10 @@
 #### Version 2015-05-26
 
+* Added label "Last updated:" (defaults to not being shown!)
+* Made separator string translateable (makes sense for a lot of languages to have a "sane default" then)
+* Added translations to the plugin; plus default .pot file and German translations
 * Added "get_the_ID()" as the default value of "post_id" parameter -- Note: in this case the Shortcode could be used as is [siu-item-updated] but within the Loop!
-* Addec CHANGES.md file for changelogs
+* Added CHANGES.md file for changelogs
 * Minor formatting stuff